Does not log all validation snapshots by default

+++ main.py
@@ -16,7 +16,7 @@ from ml4ps.reinforcement.algorithm import get_algorithm
 
 
 OmegaConf.register_new_resolver("add", lambda x, y: x+y)
-OmegaConf.register_new_resolver("mul", lambda x, y: x*y)
+OmegaConf.register_new_resolver("mul", lambda x, y: int(x*y))
 
 
 def get_vector_env(*, name, **kwargs) -> VectorEnv:
